+++ src/main/java/site/geni/FarLands/mixins/client/fluid/LavaFluidMixin.java
@@ -39,9 +39,9 @@ public abstract class LavaFluidMixin {
 	private void addParticlesProperly(World world, BlockPos blockPos, FluidState fluidState, Random random, CallbackInfo ci) {
 		if (Config.getConfig().fixParticles) {
 
-			double x = blockPos.getX() + random.nextDouble();
+			double x = blockPos.getX() + (double) random.nextFloat();
 			double y = blockPos.getY() + 1;
-			double z = blockPos.getZ() + random.nextDouble();
+			double z = blockPos.getZ() + (double) random.nextFloat();
 
 			world.addParticle(ParticleTypes.LAVA, x, y, z, 0.0D, 0.0D, 0.0D);
 			world.playSound(x, y, z, SoundEvents.BLOCK_LAVA_POP, SoundCategory.BLOCK, 0.2F + random.nextFloat() * 0.2F, 0.9F + random.nextFloat() * 0.15F, false);
